feat : alert functionality using google smpt
