--- src/main/java/com/algo/email/application/EmailService.java
@@ -1,28 +1,0 @@
-package com.algo.email.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-
-/**
- * @author : iyeong-gyo
- * @package : com.algo.email.application
- * @since : 12.03.24
- */
-@RequiredArgsConstructor
-@Service
-public class EmailService {
-
-  private final JavaMailSender emailSender;
-
-  public void sendSimpleMessage(
-      String to, String subject, String text) {
-    SimpleMailMessage message = new SimpleMailMessage();
-    message.setFrom(""); // from email
-    message.setTo(to);
-    message.setSubject(subject);
-    message.setText(text);
-    emailSender.send(message);
-  }
-}
