--- src/main/java/com/algo/email/config/MailConfig.java
@@ -1,34 +1,0 @@
-package com.algo.email.config;
-
-import java.util.Properties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-
-/**
- * @author : iyeong-gyo
- * @package : com.algo.email.config
- * @since : 12.03.24
- */
-@Configuration
-public class MailConfig {
-
-  @Bean
-  public JavaMailSender getJavaMailSender() {
-    JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
-    mailSender.setHost("smtp.gmail.com");
-    mailSender.setPort(587);
-
-    mailSender.setUsername(""); // google e-mail
-    mailSender.setPassword(""); // google app password
-
-    Properties props = mailSender.getJavaMailProperties();
-    props.put("mail.transport.protocol", "smtp");
-    props.put("mail.smtp.auth", "true");
-    props.put("mail.smtp.starttls.enable", "true");
-    props.put("mail.debug", "true");
-
-    return mailSender;
-  }
-}
