+++ src/main/resources/application-local.properties
@@ -15,7 +15,6 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 
-
 # =========================== #
 #           Http Config       #
 # =========================== #
@@ -26,3 +25,18 @@ cors.allowed-origins=http://localhost:8080/
 #   TRACE, DEBUG, INFO, WARN, ERROR #
 # ================================= #
 logging.level.root=INFO
+
+# ================================= #
+#         Email Service             #
+# ================================= #
+# sample
+spring.mail.username=sample@gmail.com
+# sample
+spring.mail.password=password
+
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.protocol=smtp
+spring.mail.default-encoding=utf-8
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
